+++ packages/build-app-helpers/src/injectTransformRuntime.ts
@@ -11,7 +11,7 @@ export default function(config) {
         const targetPlugin = formatPath('@babel/plugin-transform-runtime');
         const pluginOptions = {
           corejs: false,
-          helpers: false,
+          helpers: true,
           regenerator: true,
           useESModules: false,
         };